# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Phonebook app",
        "description": "Phonebook app",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
        "/contacts": {
            "get": {
                "tags": [
                    "Contact"
                ],
                "summary": "Get list of user's contacts",
                "operationId": "2472e81f7e9df9059ccbc6b7e8d7422a",
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfull operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Contact"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contact"
                ],
                "summary": "Create new contact",
                "operationId": "3f2942c1c536cc393583e59a9d81a5b8",
                "requestBody": {
                    "request": "Contact",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "Contact name": {
                                        "type": "string"
                                    },
                                    "Contact phone": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "200": {
                        "description": "Create new contact",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "contacts/{contact}": {
            "get": {
                "tags": [
                    "Contact"
                ],
                "summary": "Get contact info by contact_id",
                "description": "Get contact info by contact_id",
                "operationId": "f41eddd54fa9799ba3971f9863f64db2",
                "parameters": [
                    {
                        "name": "contact",
                        "in": "path",
                        "description": "contact identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieve specified contact",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contact"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "404": {
                        "description": "Contact is not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Contact"
                ],
                "description": "Update contact",
                "operationId": "be3651749745503781bd076ebdb2ea62",
                "requestBody": {
                    "request": "Contact",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Contact"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Contact"
                ],
                "summary": "Delete contact",
                "operationId": "b021fd45548ef7c9e27c0e7ac713793b",
                "parameters": [
                    {
                        "name": "contact",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete specified contact",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Cannot delete contact",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "contacts/{contact}/favourite": {
            "put": {
                "tags": [
                    "Contact"
                ],
                "description": "Mark contact as favourite",
                "operationId": "8995f549d2117fe987c93d7318fac2c7",
                "responses": {
                    "202": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Show user details",
                "operationId": "9c04d0fd3a123a8b90e19244d6e1bb75",
                "requestBody": {
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "200": {
                        "description": "Show user details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user",
                "operationId": "b4443b59cb9ea2a1e5037ad3c66d4efd",
                "requestBody": {
                    "request": "User",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "200": {
                        "description": "Update user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create new user",
                "operationId": "7b35e05a8e9f07879f2c9b53994f72ce",
                "requestBody": {
                    "request": "User",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "200": {
                        "description": "Create new user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/login": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Authenticate user",
                "operationId": "bdb29df76f33a1a3875ebdac6d8a0dea",
                "requestBody": {
                    "request": "User",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "confirm_password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/forgot-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Send reset password link for user",
                "operationId": "51b3c83fdc20cfe1fa7372e785d74b1c",
                "requestBody": {
                    "request": "User",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reset password link sent"
                    }
                }
            }
        },
        "/user/reset-password/{token}?email={email}": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Set new password for user",
                "operationId": "87cf4062a8c27aaf57142cecc63e4c7e",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "required": false,
                        "explode": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false,
                        "explode": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "request": "User",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "type": "string"
                                    },
                                    "confirm_password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Password updated"
                    },
                    "422": {
                        "description": "Unprocessable entity"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Contact": {
                "title": "Contact",
                "description": "App\\Models\\Contact",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "App\\Models\\User",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}